requirements.txt
# Dependencias del Sistema de Inventario Inteligente

# Framework web
Flask==2.3.3
Flask-CORS==4.0.0

# Base de datos
mysql-connector-python==8.1.0

# Variables de entorno
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3

# Utilidades
requests==2.31.0
Pillow==10.0.0

# Desarrollo
python-dateutil==2.8.2 